restructured the mvc model
