+++ models/items.js
@@ -1,6 +1,28 @@
 const db = require('../config/database.js');
 const { Item, User } = require('./define.js');
 
+const addNewItem = async (
+  name,
+  colour,
+  pattern,
+  weight,
+  imageURL,
+  categoryID,
+  userID
+) => {
+  console.log('addNewItem function called');
+  const create = await Item.create({
+    name: name,
+    colour: colour,
+    pattern: pattern,
+    weight: weight,
+    imageURL: imageURL,
+    categoryID: categoryID,
+    userID: userID,
+  });
+  return create;
+};
+
 // searches DB, users table for name, then returns user ID and searches items table for their items
 const selectItemsByID = async (userName) => {
   const name = await User.findOne({ where: { name: userName }, raw: true });
@@ -63,10 +85,14 @@ const selectOverallsByCatID = async (userID, catID) => {
   return sel;
 };
 
+db.addNewItem = addNewItem;
+
 db.selectItemsByID = selectItemsByID;
+
 db.selectTopsByID = selectTopsByID;
 db.selectBottomsByID = selectBottomsByID;
 db.selectOverallsByID = selectOverallsByID;
+
 db.selectTopsByCatID = selectTopsByCatID;
 db.selectBottomsByCatID = selectBottomsByCatID;
 db.selectOverallsByCatID = selectOverallsByCatID;
