+++ models/user.js
@@ -1,6 +1,13 @@
 const db = require('../config/database.js');
 const { User } = require('./define.js');
 
+const addNewUser = async (name) => {
+  const create = await User.create({
+    name: name,
+  });
+  return create;
+};
+
 const selectAllUsers = async () => {
   const all = await User.findAll({});
   return all;
@@ -18,6 +25,8 @@ const selectUserByID = async (userID) => {
   const selection = await User.findAll({ where: { id: userID }, raw: true });
   return selection;
 };
+
+db.addNewUser = addNewUser;
 
 db.selectAllUsers = selectAllUsers;
 db.selectUserByName = selectUserByName;
