+++ public/assets/js/uploadImage.js
@@ -78,8 +78,13 @@ function badUpload() {
   }, 5000);
   $('.badUpload').removeClass('displayNone');
 }
-/*On submitting the Form the checks are made to see if form is correct then posts
-to Cloudinary followed by the Database*/
+
+/*
+On submitting the Form 
+the checks are made to see if form is correct 
+then posts to Cloudinary 
+then posts to the Database
+*/
 function imageSubmit() {
   const userID = localStorage.getItem('userID');
   if (!userID) {
@@ -97,7 +102,7 @@ function imageSubmit() {
     formData.append('image', theFile[key]);
   });
   $.post({
-    url: '/upload',
+    url: '/item/addImage',
     data: formData,
     processData: false,
     contentType: false,
@@ -122,7 +127,7 @@ function imageSubmit() {
     $.ajax({
       /* Posts Entry to our Database */
       type: 'POST',
-      url: `upload/item/${userID}`,
+      url: `item/addNew/${userID}`,
       data: dataObject,
     }).then(() => {
       $('#myFile').val('');
